refactor: remove dead oauth2 code and fix build error (#231)
Remove unnecessary oauth2 transport check in FetchRepositoryContents
that was a no-op since oauth2.Transport automatically handles
authorization headers.

Also fix type mismatch in cache/storage.go DeleteByPrefix that was
preventing builds - removed unnecessary int64() cast since both
deletedCount and TotalEntries are int.

Closes #225

+++ github-profile-tools/internal/cache/storage.go
@@ -236,7 +236,7 @@ func (fs *FileStorage) DeleteByPrefix(keyPrefix string) error {
 	}
 
 	// Update stats
-	fs.stats.TotalEntries -= int64(deletedCount)
+	fs.stats.TotalEntries -= deletedCount
 	if deletedCount > 0 {
 		log.Printf("Deleted %d cache entries matching prefix: %s", deletedCount, keyPrefix)
 	}
